fix compile error 1. fix typo 2. remove enused variable 3. change some type 4. remove bad code

+++ code/mathematical_algorithms/src/segmented_sieve_of_eratosthenes/segmented_sieve_of_eratosthenes.cpp
@@ -52,7 +52,7 @@ void segmentedSieve(){
 int main(){
     segmentedSieve(); //run it to find primes..
     //Now Prime contains all the prime numbers upto segMax defined in function
-    for(int i = 0; i < Prime.size(); i++){
+    for(size_t i = 0; i < Prime.size(); i++){
         cout << Prime[i] <<" ";
     }
     //^^ Prints all the prime numbers
